Test 9 passed??

+++ Exercises/num_vector_utils.h
@@ -9,5 +9,5 @@ vector<int> only_evens(const vector<int>&);
 vector<int> only_odds(const vector<int>&);
 vector<int> nums_between(const vector<int>&, int num1, int num2);
 double num_vector_mean(const vector<int>&);
-double num_vector_median(const vector<int>&);
+float num_vector_median(const vector<int>&);
